+++ package.json
@@ -32,12 +32,16 @@
   },
   "scripts": {
     "pretest": "eslint mustache.js",
-    "test": "mocha --reporter spec",
-    "test-render": "mocha  --reporter spec test/render-test"
+    "test": "mocha --reporter spec test/*-test.js",
+    "test-render": "mocha  --reporter spec test/render-test",
+    "pre-test-browser": "node test/create-browser-suite.js",
+    "test-browser": "npm run pre-test-browser && zuul -- test/context-test.js test/parse-test.js test/scanner-test.js test/render-test-browser.js",
+    "test-browser-local": "npm run pre-test-browser && zuul --local 8080 -- test/context-test.js test/scanner-test.js test/parse-test.js test/render-test-browser.js"
   },
   "devDependencies": {
     "eslint": "^0.20.0",
-    "mocha": "~2.1.0"
+    "mocha": "~2.1.0",
+    "zuul": "^2.1.1"
   },
   "spm": {
     "main": "mustache.js",
